Changed EnemyMovement
Change EnemyMovement to use the same NavMesh system that  Troops use. Unfortunately, it suffers the same problem where some of the units get stuck on waypoints. Still thinking of ways to get around this. Maybe create some logic that unsticks them and forces them to go to the next waypoint.