Pathfinding Update
Troops now can follow different paths. Paths are groups of TroopWaypoints that the Troops follow.
Need to add the ability for the player to select which path their Troops will follow.
Need to add variables for first and last waypoint indeces. This should help with the sticking problem, as things will be more explicit.
Need to make paths wider and Troops and enemies a bit smaller.

New bug: Troops are now stopping completely after entering and exiting combat.