Modeling + Multiplayer Beginnings
Created models for basic troops and defense towers for the four playable armies: Capsule, Orb, Diamond, and Cube. Still need to implement these and create upgraded versions of everything.

Started working on adding in multiplayer. I'm probably going to gut this in favor of a lockstep framework I've found. Should also help with other areas of the game.

ToDo next:
-Offense towers for spawning troops
-Upgraded towers and troops
-Different shops for each army
-Implement new troops and towers
-Work on multiplayer
-Create scenery