FireCountdown and Gallery
Fixed a problem that caused the Fire Countdown variable for towers to only count down when the tower was targeting an enemy. It now counts down right after the tower shoots without the need to be targeting anything.

Forgot to add this for the last update: Added a Gallery mode where the player can look at the models for all of the towers and troops in the game so far. I plan to make this interactable where if you click on an object then it will display that object's data, such as its name, stats, and a short description.

Putting multiplayer functionality on the backburner for the moment to perfect the core gameplay. I will be focusing on creating a singleplayer mode where the player can play against an AI opponent. This will serve to balance things out and give me a good idea of how the multiplayer should end up working.

ToDo next:
-Create singleplayer lobby menu where the player can choose a map, an AI opponent, and change other settings.
-Create Options menu
-Add some sound and scenery
-Start on AI opponent
-Fix up the shops some more
-Add functionality to offense towers