Scenery and Music
I have added some scenery and music to the Main Menu and the Mode Select menu. I added sfx to the buttons on the main menu and plan to add sounds to all of the other buttons as well.

I will be adding music to the map as well as sfx for the turrets and troops next.