Presentation Ready Edition
Singleplayer game against an AI opponent is mostly playable. Just needs a few bugs worked out and it should be good to go and expand on.

Singleplayer Lobby menu has been added. Right now it is just a mock up of what it will look like in the future. I plan to make it so that the player can choose a race for them and the AI opponent and have that choice change the characters and scenery on screen.

TroopCombat works slightly better now, but a new bug now causes troops to blow by the waiting area and go straight into battle. On the plus side, they now move past dead enemies instead of staying in place.